week06/pyproject.toml
[tool.poetry]
name = "week06"
version = "0.1.0"
description = ""
authors = ["diligejy <sjy049@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
pymongo = "^3.11.4"
requests = "^2.25.1"
PyJWT = "^2.1.0"
python-dotenv = "^0.17.1"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]
pytest-flask = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
